move v1 onboarding to own folder
