+++ frontend/src/lib/components/Sharing/SharingModal.tsx
@@ -5,7 +5,7 @@ import { sharingLogic } from './sharingLogic'
 import { LemonButton, LemonDivider, LemonSwitch } from '@posthog/lemon-ui'
 import { copyToClipboard } from 'lib/utils'
 import { IconGlobeLock, IconInfo, IconLink, IconLock, IconUnfoldLess, IconUnfoldMore } from '../icons'
-import { CodeSnippet, Language } from 'scenes/ingestion/frameworks/CodeSnippet'
+import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
 import { DashboardCollaboration } from 'scenes/dashboard/DashboardCollaborators'
 import { Field } from 'lib/forms/Field'
 import { Tooltip } from 'lib/components/Tooltip'
