+++ frontend/src/scenes/apps/ErrorDetailsModal.tsx
@@ -7,7 +7,7 @@ import { LemonLabel } from 'lib/components/LemonLabel/LemonLabel'
 import { LemonButton } from 'lib/components/LemonButton'
 import { IconChevronLeft, IconChevronRight, IconUnfoldLess, IconUnfoldMore } from 'lib/components/icons'
 import { LemonModal } from 'lib/components/LemonModal'
-import { CodeSnippet, Language } from '../ingestion/frameworks/CodeSnippet'
+import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
 
 export function ErrorDetailsModal(): JSX.Element {
     const { errorDetails, errorDetailsModalError, errorDetailsLoading } = useValues(appMetricsSceneLogic)
