+++ frontend/src/scenes/authentication/PasswordReset.tsx
@@ -3,7 +3,7 @@ Scene to request a password reset email.
 */
 import { useActions, useValues } from 'kea'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
-import { CodeSnippet, Language } from 'scenes/ingestion/frameworks/CodeSnippet'
+import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
 import { passwordResetLogic } from './passwordResetLogic'
 import { router } from 'kea-router'
 import { SceneExport } from 'scenes/sceneTypes'
