+++ frontend/src/scenes/experiments/ExperimentCodeSnippets.tsx
@@ -1,5 +1,5 @@
 import { Link } from '@posthog/lemon-ui'
-import { CodeSnippet, Language } from 'scenes/ingestion/frameworks/CodeSnippet'
+import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
 
 function ServerSideWarning(): JSX.Element {
     return (
