+++ frontend/src/scenes/ingestion/v1/IngestionV1.stories.tsx
@@ -1,10 +1,10 @@
 import { Meta } from '@storybook/react'
 
-import { IngestionWizard } from './IngestionWizard'
+import { IngestionWizardV1 } from './IngestionWizard'
 
 export default {
     title: 'Scenes-Other/Onboarding',
     parameters: { layout: 'fullscreen', options: { showPanel: false }, viewMode: 'story' },
 } as Meta
 
-export const Ingestion = (): JSX.Element => <IngestionWizard />
+export const IngestionV1 = (): JSX.Element => <IngestionWizardV1 />
