+++ frontend/src/scenes/ingestion/v1/IngestionWizard.scss
