+++ frontend/src/scenes/ingestion/v1/IngestionWizard.tsx
@@ -1,13 +1,13 @@
 import { useEffect } from 'react'
 import './IngestionWizard.scss'
 
-import { VerificationPanel } from 'scenes/ingestion/panels/VerificationPanel'
-import { InstructionsPanel } from 'scenes/ingestion/panels/InstructionsPanel'
-import { MOBILE, BACKEND, WEB, BOOKMARKLET, THIRD_PARTY } from 'scenes/ingestion/constants'
+import { VerificationPanel } from 'scenes/ingestion/v1/panels/VerificationPanel'
+import { InstructionsPanel } from 'scenes/ingestion/v1/panels/InstructionsPanel'
+import { MOBILE, BACKEND, WEB, BOOKMARKLET, THIRD_PARTY } from 'scenes/ingestion/v1/constants'
 import { useValues, useActions } from 'kea'
-import { ingestionLogic } from 'scenes/ingestion/ingestionLogic'
-import { FrameworkPanel } from 'scenes/ingestion/panels/FrameworkPanel'
-import { PlatformPanel } from 'scenes/ingestion/panels/PlatformPanel'
+import { ingestionLogic } from 'scenes/ingestion/v1/ingestionLogic'
+import { FrameworkPanel } from 'scenes/ingestion/v1/panels/FrameworkPanel'
+import { PlatformPanel } from 'scenes/ingestion/v1/panels/PlatformPanel'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { BookmarkletPanel } from './panels/BookmarkletPanel'
 import { ThirdPartyPanel } from './panels/ThirdPartyPanel'
