+++ frontend/src/scenes/ingestion/v1/frameworks/ElixirInstructions.tsx
@@ -1,4 +1,4 @@
-import { CodeSnippet, Language } from './CodeSnippet'
+import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
 import { useValues } from 'kea'
 import { teamLogic } from 'scenes/teamLogic'
