+++ frontend/src/scenes/ingestion/v1/panels/BookmarkletPanel.tsx
