+++ frontend/src/scenes/ingestion/v1/panels/FrameworkPanel.tsx
@@ -1,7 +1,7 @@
 import { useActions, useValues } from 'kea'
-import { CardContainer } from 'scenes/ingestion/CardContainer'
-import { ingestionLogic } from 'scenes/ingestion/ingestionLogic'
-import { API, mobileFrameworks, BACKEND, webFrameworks } from 'scenes/ingestion/constants'
+import { CardContainer } from 'scenes/ingestion/v1/CardContainer'
+import { ingestionLogic } from 'scenes/ingestion/v1/ingestionLogic'
+import { API, mobileFrameworks, BACKEND, webFrameworks } from 'scenes/ingestion/v1/constants'
 import { LemonButton } from 'lib/components/LemonButton'
 import './Panels.scss'
