+++ frontend/src/scenes/ingestion/v1/panels/InstructionsPanel.scss
