+++ frontend/src/scenes/ingestion/v1/panels/Panels.scss
