+++ frontend/src/scenes/ingestion/v1/panels/ThirdPartyIcons.tsx
