+++ frontend/src/scenes/ingestion/v1/panels/ThirdPartyPanel.tsx
@@ -9,7 +9,7 @@ import { IconOpenInNew } from 'lib/components/icons'
 import { PluginDrawer } from 'scenes/plugins/edit/PluginDrawer'
 import { pluginsLogic } from 'scenes/plugins/pluginsLogic'
 import { PluginInstallationType, PluginRepositoryEntryType, PluginTypeWithConfig } from 'scenes/plugins/types'
-import { CodeSnippet } from '../frameworks/CodeSnippet'
+import { CodeSnippet } from 'lib/components/CodeSnippet'
 import { teamLogic } from 'scenes/teamLogic'
 import { LemonTag } from 'lib/components/LemonTag/LemonTag'
 import { Link } from 'lib/components/Link'
