+++ frontend/src/scenes/ingestion/v1/panels/VerificationPanel.tsx
@@ -1,7 +1,7 @@
 import { useActions, useValues } from 'kea'
 import { useInterval } from 'lib/hooks/useInterval'
-import { CardContainer } from 'scenes/ingestion/CardContainer'
-import { ingestionLogic } from 'scenes/ingestion/ingestionLogic'
+import { CardContainer } from 'scenes/ingestion/v1/CardContainer'
+import { ingestionLogic } from 'scenes/ingestion/v1/ingestionLogic'
 import { teamLogic } from 'scenes/teamLogic'
 import { Spinner } from 'lib/components/Spinner/Spinner'
 import { LemonButton } from 'lib/components/LemonButton'
