+++ frontend/src/scenes/ingestion/v2/panels/FrameworkPanel.tsx
@@ -1,7 +1,7 @@
 import { useActions, useValues } from 'kea'
-import { CardContainer } from 'scenes/ingestion/CardContainer'
+import { CardContainer } from 'scenes/ingestion/v2/CardContainer'
 import { ingestionLogic } from '../ingestionLogic'
-import { API, mobileFrameworks, BACKEND, webFrameworks } from 'scenes/ingestion/constants'
+import { API, mobileFrameworks, BACKEND, webFrameworks } from 'scenes/ingestion/v2/constants'
 import { LemonButton } from 'lib/components/LemonButton'
 import './Panels.scss'
 import { IngestionInviteMembersButton } from '../IngestionInviteMembersButton'
