+++ frontend/src/scenes/plugins/plugin/PluginError.tsx
@@ -1,7 +1,7 @@
 import { Button, Popover, Tag } from 'antd'
 import { ClearOutlined } from '@ant-design/icons'
 import { PluginErrorType } from '~/types'
-import { CodeSnippet, Language } from 'scenes/ingestion/frameworks/CodeSnippet'
+import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
 import { dayjs } from 'lib/dayjs'
 
 export function PluginError({ error, reset }: { error: PluginErrorType; reset?: () => void }): JSX.Element | null {
