+++ frontend/src/scenes/project/Settings/SlackIntegration.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { useActions, useValues } from 'kea'
 import { getSlackAppManifest, integrationsLogic } from './integrationsLogic'
-import { CodeSnippet, Language } from 'scenes/ingestion/frameworks/CodeSnippet'
+import { CodeSnippet, Language } from 'lib/components/CodeSnippet'
 import { LemonButton, Link } from '@posthog/lemon-ui'
 import { IconDelete, IconSlack } from 'lib/components/icons'
 import { UserActivityIndicator } from 'lib/components/UserActivityIndicator/UserActivityIndicator'
