+++ frontend/src/scenes/project/Settings/index.tsx
@@ -7,7 +7,7 @@ import { WebhookIntegration } from './WebhookIntegration'
 import { useAnchor } from 'lib/hooks/useAnchor'
 import { router } from 'kea-router'
 import { ToolbarSettings } from './ToolbarSettings'
-import { CodeSnippet } from 'scenes/ingestion/frameworks/CodeSnippet'
+import { CodeSnippet } from 'lib/components/CodeSnippet'
 import { teamLogic } from 'scenes/teamLogic'
 import { DangerZone } from './DangerZone'
 import { PageHeader } from 'lib/components/PageHeader'
